Base-stage samples + remaining Stage-4/5/6/8 scripts + standalone section draft
- Base OLMo-2-1124-7B samples: 2000 (alpacaeval) + 1000 (nbcurated) rows, K=10.
- 4_score_baselines.py: EAD + distinct-n (Kirk's formulas) + SentBERT-sim.
- 5_analyze_and_figures.py: paired Wilcoxons with Bonferroni, Cohen's d_z,
  figures, paper_macros_rlhf.tex, tables/rlhf_diversity.tex.
- 6_upload_to_hf.py: private HF dataset upload (AMindToThink/ prefix).
- paper/rlhf_experiment.tex: standalone section draft with \providecommand
  fallbacks so it compiles before analysis runs.
- vLLM dtype fix: Quadro RTX 8000 is CC 7.5 → fp16 (no bf16 support).
- sentence-transformers added as optional dep.

+++ scripts/rlhf_experiment/2_sample_responses.py
@@ -112,10 +112,11 @@ def sample_stage(stage: str, max_new_tokens: int, seed: int) -> None:
     # Tokeniser is needed for chat templating on the non-base stages.
     tokenizer = AutoTokenizer.from_pretrained(hf_path)
 
-    # vLLM engine
+    # vLLM engine.
+    # Quadro RTX 8000 is compute-capability 7.5 → no bf16 support; use fp16.
     llm = LLM(
         model=hf_path,
-        dtype="bfloat16",
+        dtype="float16",
         trust_remote_code=True,
         gpu_memory_utilization=0.85,
         max_model_len=2048,
